ContainerファイルをRootNavContentにインライン化
Containerを経由せず他の画面と同様にRootNavContent内で
直接ViewModelを生成・Screenを呼び出す形に統一した
